requirements.txt
# requirements.txt for MRVG
accelerate==0.19.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
av==10.0.0
contourpy==1.3.2
cycler==0.12.1
decord==0.6.0
defusedxml==0.7.1
diffusers==0.16.0
einops==0.6.1
ffmpeg==1.4
fonttools==4.58.5
fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.19.4
imageio==2.31.1
imageio-ffmpeg==0.4.9
importlib-metadata==8.7.0
kiwisolver==1.4.8
matplotlib==3.10.3
mypy-extensions==1.1.0
natsort==8.4.0
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.0.1
platformdirs==4.3.8
psutil==7.0.0
pycocotools==2.0.10
pyparsing==3.2.3
pyre-extensions==0.0.29
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
rotary-embedding-torch==0.3.6
scipy==1.16.0
six==1.17.0
supervision==0.25.1
timm==0.6.13
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.28.1
typing-inspect==0.9.0
tzdata==2025.2
xformers==0.0.20
yapf==0.43.0
zipp==3.23.0
